remove sourceformlogic and circular call
