+++ frontend/src/scenes/data-warehouse/external/forms/SourceForm.tsx
@@ -4,8 +4,7 @@ import { LemonField } from 'lib/lemon-ui/LemonField'
 
 import { SourceConfig } from '~/types'
 
-import { SOURCE_DETAILS } from '../../new/sourceWizardLogic'
-import { sourceFormLogic } from './sourceFormLogic'
+import { SOURCE_DETAILS, sourceWizardLogic } from '../../new/sourceWizardLogic'
 
 interface SourceFormProps {
     sourceConfig: SourceConfig
@@ -14,8 +13,8 @@ interface SourceFormProps {
 export default function SourceForm({ sourceConfig }: SourceFormProps): JSX.Element {
     return (
         <Form
-            logic={sourceFormLogic}
-            props={{ sourceConfig }}
+            logic={sourceWizardLogic}
+            props={{}}
             formKey="sourceConnectionDetails"
             className="space-y-4"
             enableFormOnSubmit
