+++ frontend/src/scenes/data-warehouse/new/NewSourceWizard.tsx
@@ -1,5 +1,6 @@
 import { LemonButton } from '@posthog/lemon-ui'
 import { useActions, useValues } from 'kea'
+import { router } from 'kea-router'
 import { PageHeader } from 'lib/components/PageHeader'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
@@ -25,7 +26,7 @@ export const scene: SceneExport = {
 }
 export function NewSourceWizard(): JSX.Element {
     const { modalTitle, modalCaption } = useValues(sourceWizardLogic)
-    const { onBack, onSubmit, closeWizard } = useActions(sourceWizardLogic)
+    const { onBack, onSubmit, closeWizard, cancelWizard } = useActions(sourceWizardLogic)
     const { currentStep, isLoading, canGoBack, canGoNext, nextButtonText, showSkipButton } =
         useValues(sourceWizardLogic)
 
@@ -64,17 +65,17 @@ export function NewSourceWizard(): JSX.Element {
         )
     }, [currentStep, isLoading, canGoNext, canGoBack, nextButtonText, showSkipButton])
 
+    const onCancel = (): void => {
+        cancelWizard()
+        router.actions.push(urls.dataWarehouse())
+    }
+
     return (
         <>
             <PageHeader
                 buttons={
                     <>
-                        <LemonButton
-                            type="secondary"
-                            center
-                            data-attr="source-form-cancel-button"
-                            to={urls.dataWarehouse()}
-                        >
+                        <LemonButton type="secondary" center data-attr="source-form-cancel-button" onClick={onCancel}>
                             Cancel
                         </LemonButton>
                     </>
