+++ frontend/src/scenes/data-warehouse/redirect/DataWarehouseRedirectScene.tsx
@@ -1,32 +1,17 @@
-import { LemonButton, LemonInput } from '@posthog/lemon-ui'
-import { Form } from 'kea-forms'
-import { LemonField } from 'lib/lemon-ui/LemonField'
-import { sourceFormLogic } from 'scenes/data-warehouse/external/forms/sourceFormLogic'
+import { Spinner } from '@posthog/lemon-ui'
 import { SceneExport } from 'scenes/sceneTypes'
+
+import { sourceWizardLogic } from '../new/sourceWizardLogic'
 
 export const scene: SceneExport = {
     component: DataWarehouseRedirectScene,
-    logic: sourceFormLogic,
+    logic: sourceWizardLogic,
 }
 
 export function DataWarehouseRedirectScene(): JSX.Element {
     return (
         <div className="text-left flex flex-col">
-            <h2>Configure</h2>
-            <p>Add a prefix to your tables to avoid conflicts with other data sources</p>
-            <Form
-                logic={sourceFormLogic}
-                formKey="externalDataSource"
-                className="space-y-4 max-w-100"
-                enableFormOnSubmit
-            >
-                <LemonField name="prefix" label="Table prefix">
-                    <LemonInput />
-                </LemonField>
-                <LemonButton type="primary" htmlType="submit">
-                    Submit
-                </LemonButton>
-            </Form>
+            <Spinner />
         </div>
     )
 }
